requirements.txt
pyyaml==3.12
attrs==18.1
